Implement OverlayTabControl class
An abstraction for OverlayHeaderTabControl
