+++ osu.Game/Overlays/OverlayHeaderTabControl.cs
@@ -1,153 +1,28 @@
 ﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Shapes;
+using osu.Framework.Allocation;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Events;
 using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Graphics.UserInterface;
-using osuTK;
-using osuTK.Graphics;
 
 namespace osu.Game.Overlays
 {
-    public class OverlayHeaderTabControl : TabControl<string>
+    public class OverlayHeaderTabControl : OverlayTabControl<string>
     {
-        private readonly Box bar;
+        protected override TabItem<string> CreateTabItem(string value) => new OverlayHeaderTabItem(value);
 
-        private Color4 accentColour = Color4.White;
-
-        public Color4 AccentColour
+        [BackgroundDependencyLoader]
+        private void load(OsuColour colours)
         {
-            get => accentColour;
-            set
-            {
-                if (accentColour == value)
-                    return;
-
-                accentColour = value;
-                bar.Colour = value;
-
-                foreach (TabItem<string> tabItem in TabContainer)
-                {
-                    ((HeaderTabItem)tabItem).AccentColour = value;
-                }
-            }
+            AccentColour = colours.Seafoam;
         }
 
-        public new MarginPadding Padding
+        private class OverlayHeaderTabItem : OverlayTabItem<string>
         {
-            get => TabContainer.Padding;
-            set => TabContainer.Padding = value;
-        }
-
-        public OverlayHeaderTabControl()
-        {
-            TabContainer.Masking = false;
-            TabContainer.Spacing = new Vector2(15, 0);
-
-            AddInternal(bar = new Box
-            {
-                RelativeSizeAxes = Axes.X,
-                Height = 2,
-                Anchor = Anchor.BottomLeft,
-                Origin = Anchor.CentreLeft
-            });
-        }
-
-        protected override Dropdown<string> CreateDropdown() => null;
-
-        protected override TabItem<string> CreateTabItem(string value) => new HeaderTabItem(value)
-        {
-            AccentColour = AccentColour
-        };
-
-        private class HeaderTabItem : TabItem<string>
-        {
-            private readonly OsuSpriteText text;
-            private readonly ExpandingBar bar;
-
-            private Color4 accentColour;
-
-            public Color4 AccentColour
-            {
-                get => accentColour;
-                set
-                {
-                    if (accentColour == value)
-                        return;
-
-                    accentColour = value;
-                    bar.Colour = value;
-
-                    updateState();
-                }
-            }
-
-            public HeaderTabItem(string value)
+            public OverlayHeaderTabItem(string value)
                 : base(value)
             {
-                AutoSizeAxes = Axes.X;
-                RelativeSizeAxes = Axes.Y;
-
-                Children = new Drawable[]
-                {
-                    text = new OsuSpriteText
-                    {
-                        Margin = new MarginPadding { Bottom = 10 },
-                        Origin = Anchor.BottomLeft,
-                        Anchor = Anchor.BottomLeft,
-                        Text = value,
-                        Font = OsuFont.GetFont()
-                    },
-                    bar = new ExpandingBar
-                    {
-                        Anchor = Anchor.BottomCentre,
-                        ExpandedSize = 7.5f,
-                        CollapsedSize = 0
-                    },
-                    new HoverClickSounds()
-                };
-            }
-
-            protected override bool OnHover(HoverEvent e)
-            {
-                base.OnHover(e);
-
-                updateState();
-
-                return true;
-            }
-
-            protected override void OnHoverLost(HoverLostEvent e)
-            {
-                base.OnHoverLost(e);
-
-                updateState();
-            }
-
-            protected override void OnActivated() => updateState();
-
-            protected override void OnDeactivated() => updateState();
-
-            private void updateState()
-            {
-                if (Active.Value || IsHovered)
-                {
-                    text.FadeColour(Color4.White, 120, Easing.InQuad);
-                    bar.Expand();
-
-                    if (Active.Value)
-                        text.Font = text.Font.With(weight: FontWeight.Bold);
-                }
-                else
-                {
-                    text.FadeColour(AccentColour, 120, Easing.InQuad);
-                    bar.Collapse();
-                    text.Font = text.Font.With(weight: FontWeight.Medium);
-                }
+                Text.Text = value;
             }
         }
     }
